calculate yearly ssavings

+++ python_introduction/finance_calculator.py
@@ -2,8 +2,8 @@ import locale
 
 locale.setlocale(locale.LC_ALL, '')
 
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 monthly_savings = monthly_income - monthly_expenses
 projected_1year_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
